# Documents/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Farm API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Sensors": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Get list of Sensors",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SensorDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SensorDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SensorDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sensors"
        ],
        "summary": "Add a Sensor",
        "requestBody": {
          "description": "Sensor object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorAddDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorAddDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SensorAddDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sensors/{id}": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Get Sensor based on Sensor ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "sensor ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "sensor ID",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sensors/{sensorId}": {
      "put": {
        "tags": [
          "Sensors"
        ],
        "summary": "Update a Sensor",
        "parameters": [
          {
            "name": "sensorId",
            "in": "path",
            "description": "Sensor ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Sensor ID",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "State dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sensors/average": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Get Sensor Average per month",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Sensor state like Inventory, Deployed, FarmerTriage, Returned, Dead, Refurbished",
            "schema": {
              "type": "string",
              "description": "Sensor state like Inventory, Deployed, FarmerTriage, Returned, Dead, Refurbished",
              "nullable": true
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "Year",
            "schema": {
              "type": "integer",
              "description": "Year",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AveragePerMonthDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AveragePerMonthDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AveragePerMonthDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sensors/count": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Get Sensor Count per month",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Sensor state like Inventory, Deployed, FarmerTriage, Returned, Dead, Refurbished",
            "schema": {
              "type": "string",
              "description": "Sensor state like Inventory, Deployed, FarmerTriage, Returned, Dead, Refurbished",
              "nullable": true
            }
          },
          {
            "name": "month",
            "in": "query",
            "description": "Month like January ,February ,March ,April ,May ,June ,July ,August ,September ,October ,November ,December",
            "schema": {
              "type": "string",
              "description": "Month like January ,February ,March ,April ,May ,June ,July ,August ,September ,October ,November ,December",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "StateDto": {
        "required": [
          "state"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SensorDto": {
        "type": "object",
        "properties": {
          "sensorId": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "farm": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SensorAddDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "nullable": true
          },
          "farm": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AveragePerMonthDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}